+++ next-docs/pages/components/charts/geoMapChart.tsx
@@ -4,12 +4,12 @@ import Preview from '../../../components/codePreview/Preview';
 import CodeSnippet from '../../../components/CodeSnippet';
 import Table from '../../../components/Table';
 
-import Default from '../../../public/examples/charts/geomapchart/Default';
+import Default from '../../../public/examples/charts/geoMapChart/Default';
 
 import useExamples from '../../../utils/useExamples';
 
 export default function PageGeoMapChart() {
-  const examples = useExamples('charts/GeoMapChart');
+  const examples = useExamples('charts/geoMapChart');
 
   return (
     <>