+++ next-docs/pages/components/charts/tableChart.tsx
@@ -5,14 +5,14 @@ import Preview from '../../../components/codePreview/Preview';
 import Table from '../../../components/Table';
 import CodeSnippet from '../../../components/CodeSnippet';
 
-import Default from '../../../public/examples/charts/tablechart/Default';
-import Loading from '../../../public/examples/charts/tablechart/Loading';
-import Active from '../../../public/examples/charts/tablechart/Active';
+import Default from '../../../public/examples/charts/tableChart/Default';
+import Loading from '../../../public/examples/charts/tableChart/Loading';
+import Active from '../../../public/examples/charts/tableChart/Active';
 
 import useExamples from '../../../utils/useExamples';
 
 export default function PageLineChart() {
-  const examples = useExamples('charts/TableChart');
+  const examples = useExamples('charts/tableChart');
 
   return (
     <>