+++ next-docs/pages/components/progress/linear.tsx
@@ -7,7 +7,7 @@ import Default from '../../../public/examples/progress/linear/Default';
 import useExamples from '../../../utils/useExamples';
 
 export default function PageProgressLinear() {
-  const examples = useExamples('progress/Linear');
+  const examples = useExamples('progress/linear');
 
   return (
     <>