+++ next-docs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 7.12.0
+
+### Minor Changes
+
+- Fix examples
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-charts@7.12.0
+  - @heathmont/moon-draggabletable@7.12.0
+  - @heathmont/moon-accordion@7.12.0
+  - @heathmont/moon-assets@7.12.0
+  - @heathmont/moon-components@7.12.0
+  - @heathmont/moon-core@7.12.0
+  - @heathmont/moon-icons@7.12.0
+  - @heathmont/moon-pagination@7.12.0
+  - @heathmont/moon-popover@7.12.0
+  - @heathmont/moon-table@7.12.0
+  - @heathmont/moon-themes@7.12.0
+  - @heathmont/moon-utils@7.12.0
+
 ## 7.11.1
 
 ### Patch Changes