+++ packages/draggabletable/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 7.12.0
+
+### Minor Changes
+
+- Fix examples
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-components@7.12.0
+  - @heathmont/moon-table@7.12.0
+  - @heathmont/moon-themes@7.12.0
+  - @heathmont/moon-utils@7.12.0
+
 ## 7.11.1
 
 ### Patch Changes